Move all config tests into single file
